+++ CHANGELOG.md
@@ -13,5 +13,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### Fixes
 
 - Issue with the bundled JRE showing prompt that it's not managed by the launcher when it is
+- Not giving the option to install bundled JRE when using out of date Java and on a force update
 
 ### Misc